refactor long main file to 4 handlers
